feat: refactor  AssetRegistryClient useAssetColumns
